[AIRFLOW-6543][AIP-21] Promotion of contrib classes to the core (#7145)
* [AIP-21] Move contrib.hooks.fs_hook hooks.filesystem

* [AIP-21] Move contrib.sensors.weekday_sensor sensors.weekday_sensor

* [AIP-21] Move contrib.sensors.file_sensor sensors.filesystem
