+++ airflow/contrib/hooks/fs_hook.py
@@ -16,31 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-#
+"""This module is deprecated. Please use `airflow.hooks.filesystem`."""
 
-from airflow.hooks.base_hook import BaseHook
+import warnings
 
+# pylint: disable=unused-import
+from airflow.hooks.filesystem import FSHook  # noqa
 
-class FSHook(BaseHook):
-    """
-    Allows for interaction with an file server.
-
-    Connection should have a name and a path specified under extra:
-
-    example:
-    Conn Id: fs_test
-    Conn Type: File (path)
-    Host, Schema, Login, Password, Port: empty
-    Extra: {"path": "/tmp"}
-    """
-
-    def __init__(self, conn_id='fs_default'):
-        conn = self.get_connection(conn_id)
-        self.basepath = conn.extra_dejson.get('path', '')
-        self.conn = conn
-
-    def get_conn(self):
-        pass
-
-    def get_path(self):
-        return self.basepath
+warnings.warn(
+    "This module is deprecated. Please use `airflow.hooks.filesystem`.",
+    DeprecationWarning, stacklevel=2
+)
