+++ tests/sensors/test_weekday_sensor.py
@@ -21,10 +21,10 @@
 import unittest
 
 from airflow import DAG, models
-from airflow.contrib.sensors.weekday_sensor import DayOfWeekSensor
 from airflow.contrib.utils.weekday import WeekDay
 from airflow.exceptions import AirflowSensorTimeout
 from airflow.models import DagBag, TaskFail
+from airflow.sensors.weekday_sensor import DayOfWeekSensor
 from airflow.settings import Session
 from airflow.utils.timezone import datetime
